pyproject.toml
[project]
name = "yes"
version = "0.1.0"
description = ""
authors = [
    {name = "kirillb", email = "slrill97@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "uvicorn (>=0.29.0,<1.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "wb_front"
version = "0.1.0"
description = "..."
authors = ["..."]
packages = [{ include = "app" }]

[tool.poetry.scripts]
start = "start:main"

[tool.poetry.dependencies]
nodejs = "*"